+++ lib/ndr_import/file/base.rb
@@ -9,12 +9,29 @@ module NdrImport
   module File
     # All common base file handler logic is defined here.
     class Base
-      def initialize(filename, format, options = {})
-        @filename = filename
+      # Until subclasses explicitly override this method, it is assumed that they cannot deal
+      # with IO streams instead of filenames. Once all of them can, this could be removed,
+      # but with Zip files needing an unzip_path, this is currently unlikely.
+      def self.can_stream_data?
+        false
+      end
+
+      def initialize(filename_or_stream, format, options = {})
+        # Set different variable depending on the type of source
         @format = format
         @options = options.stringify_keys
 
-        validate_filename_is_safe_and_readable!
+        if stream?(filename_or_stream)
+          unless self.class.can_stream_data?
+            raise ArgumentError,
+                  "#{filename_or_stream.class.name} is not supported"
+          end
+
+          @stream = filename_or_stream
+        else
+          @filename = filename_or_stream
+          validate_filename_is_safe_and_readable!
+        end
       end
 
       # This method iterates over the files in the given file and yields the filenames.
@@ -62,7 +79,7 @@ module NdrImport
       # end
       #
       def rows
-        fail "Implement #{self.class}#rows"
+        raise "Implement #{self.class}#rows"
       end
 
       def validate_filename_is_safe_and_readable!
@@ -71,6 +88,20 @@ module NdrImport
         # Ensure that we're allowed to read from the safe path:
         # (they can be configured to be write-only, for example)
         SafeFile.verify_mode(@filename, 'r')
+      end
+
+      # This method determines if the given object is a stream. Used to distingush between
+      # the existing behavour of exclusively using String filenames.
+      def stream?(object)
+        object.is_a?(IO) || object.is_a?(StringIO)
+      end
+
+      # This is a helper method that returns the object if it is an IO stream,
+      # otherwise it will return a SafePath of the file path.
+      def safepath_or_stream(object)
+        return object if stream?(object)
+
+        SafeFile.safepath_to_string(object)
       end
     end
   end