+++ test/file/acro_form_test.rb
@@ -34,6 +34,10 @@ module NdrImport
           assert_equal expected_row, sheet.to_a.first
         end
       end
+
+      test 'should declare that it does not handle IO streams' do
+        refute NdrImport::File::AcroForm.can_stream_data?
+      end
     end
   end
 end