+++ test/file/pdf_test.rb
@@ -31,6 +31,10 @@ module NdrImport
           end
         end
       end
+
+      test 'should declare that it does not handle IO streams' do
+        refute NdrImport::File::Pdf.can_stream_data?
+      end
     end
   end
 end