+++ test/file/word_test.rb
@@ -30,6 +30,10 @@ module NdrImport
           end
         end
       end
+
+      test 'should declare that it does not handle IO streams' do
+        refute NdrImport::File::Word.can_stream_data?
+      end
     end
   end
 end